reorg transaction folder structure
