+++ crates/cfxcore/core/src/transaction_pool/mod.rs
@@ -2,22 +2,25 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-mod account_cache;
+mod deferred_pool;
 mod error;
 mod garbage_collector;
-mod nonce_pool;
+mod pool_metrics;
+mod state_provider;
 mod transaction_pool_inner;
 
-pub use self::transaction_pool_inner::TransactionStatus;
+pub use error::TransactionPoolError;
+pub use transaction_pool_inner::TransactionStatus;
+
 use crate::{
     block_data_manager::BlockDataManager,
     consensus::BestInformation,
     transaction_pool::{
-        nonce_pool::TxWithReadyInfo, transaction_pool_inner::PendingReason,
+        deferred_pool::TxWithReadyInfo, pool_metrics::*,
+        transaction_pool_inner::PendingReason,
     },
     verification::{VerificationConfig, VerifyTxLocalMode, VerifyTxMode},
 };
-use account_cache::AccountCache;
 use cfx_executor::{
     machine::Machine, spec::TransitionsEpochHeight, state::State,
 };
@@ -35,18 +38,16 @@ use cfx_types::{
     AddressWithSpace as Address, AllChainID, Space, SpaceMap, H256, U256,
 };
 use cfx_vm_types::Spec;
-pub use error::TransactionPoolError;
+
 use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
-use metrics::{
-    register_meter_with_group, Gauge, GaugeUsize, Lock, Meter, MeterTimer,
-    RwLockExtensions,
-};
+use metrics::{MeterTimer, RwLockExtensions};
 use parking_lot::{Mutex, RwLock};
 use primitives::{
     block::BlockHeight,
     block_header::{compute_next_price, compute_next_price_tuple},
     Account, SignedTransaction, Transaction, TransactionWithSignature,
 };
+use state_provider::StateProvider;
 use std::{
     cmp::{max, min},
     collections::{hash_map::HashMap, BTreeSet},
@@ -58,39 +59,6 @@ use std::{
     },
 };
 use transaction_pool_inner::TransactionPoolInner;
-
-lazy_static! {
-    static ref TX_POOL_DEFERRED_GAUGE: Arc<dyn Gauge<usize>> =
-        GaugeUsize::register_with_group("txpool", "stat_deferred_txs");
-    static ref TX_POOL_UNPACKED_GAUGE: Arc<dyn Gauge<usize>> =
-        GaugeUsize::register_with_group("txpool", "stat_unpacked_txs");
-    static ref TX_POOL_READY_GAUGE: Arc<dyn Gauge<usize>> =
-        GaugeUsize::register_with_group("txpool", "stat_ready_accounts");
-    static ref INSERT_TPS: Arc<dyn Meter> =
-        register_meter_with_group("txpool", "insert_tps");
-    static ref INSERT_TXS_TPS: Arc<dyn Meter> =
-        register_meter_with_group("txpool", "insert_txs_tps");
-    static ref INSERT_TXS_SUCCESS_TPS: Arc<dyn Meter> =
-        register_meter_with_group("txpool", "insert_txs_success_tps");
-    static ref INSERT_TXS_FAILURE_TPS: Arc<dyn Meter> =
-        register_meter_with_group("txpool", "insert_txs_failure_tps");
-    static ref TX_POOL_INSERT_TIMER: Arc<dyn Meter> =
-        register_meter_with_group("timer", "tx_pool::insert_new_tx");
-    static ref TX_POOL_VERIFY_TIMER: Arc<dyn Meter> =
-        register_meter_with_group("timer", "tx_pool::verify");
-    static ref TX_POOL_GET_STATE_TIMER: Arc<dyn Meter> =
-        register_meter_with_group("timer", "tx_pool::get_state");
-    static ref INSERT_TXS_QUOTA_LOCK: Lock =
-        Lock::register("txpool_insert_txs_quota_lock");
-    static ref INSERT_TXS_ENQUEUE_LOCK: Lock =
-        Lock::register("txpool_insert_txs_enqueue_lock");
-    static ref PACK_TRANSACTION_LOCK: Lock =
-        Lock::register("txpool_pack_transactions");
-    static ref NOTIFY_BEST_INFO_LOCK: Lock =
-        Lock::register("txpool_notify_best_info");
-    static ref NOTIFY_MODIFIED_LOCK: Lock =
-        Lock::register("txpool_notify_modified_info");
-}
 
 pub struct TxPoolConfig {
     pub capacity: usize,
@@ -689,7 +657,7 @@ impl TransactionPool {
     // the packed tag provided
     // if force tag is true, the replacement in nonce pool must be happened
     pub fn add_transaction_with_readiness_check(
-        &self, inner: &mut TransactionPoolInner, account_cache: &AccountCache,
+        &self, inner: &mut TransactionPoolInner, account_cache: &StateProvider,
         transaction: Arc<SignedTransaction>, packed: bool, force: bool,
     ) -> Result<(), TransactionPoolError> {
         inner.insert_transaction_with_readiness_check(
@@ -1217,9 +1185,9 @@ impl TransactionPool {
         Ok(())
     }
 
-    fn get_best_state_account_cache(&self) -> AccountCache {
+    fn get_best_state_account_cache(&self) -> StateProvider {
         let _timer = MeterTimer::time_func(TX_POOL_GET_STATE_TIMER.as_ref());
-        AccountCache::new((self.best_executed_state.lock()).clone())
+        StateProvider::new((self.best_executed_state.lock()).clone())
     }
 
     pub fn ready_for_mining(&self) -> bool {
